+++ src/pytack/__init__.py
@@ -1,10 +1,10 @@
-"""__init__ for pkgdevx."""
+"""__init__ for pytack."""
 
 import contextlib
 from importlib import metadata
 from pathlib import Path
-from pkgdevx import exceptions, standards
-from pkgdevx.logging import configure_pkg_logging
+from pytack import exceptions, standards
+from pytack.logging import configure_pkg_logging
 
 __all__: list[str] = ["exceptions", "standards"]
 
@@ -13,4 +13,4 @@ PKGDEV_ROOT: Path = Path(__file__).parent
 configure_pkg_logging()
 
 with contextlib.suppress(metadata.PackageNotFoundError):
-    __version__: str = metadata.version("pkgdevx")
+    __version__: str = metadata.version("pytack")